facerecognition/backend/requirements.txt
# Backend Server dependencies
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
websockets>=12.0
insightface>=0.7.3
onnxruntime-gpu>=1.17.0
opencv-python>=4.9.0.80
numpy>=1.24.0
